python/requirements.txt
Flask
requests
psycopg2-binary
prometheus_client
prometheus-flask-exporter
